+++ tests/wdio/secondaryNavigation-spec.js
@@ -21,32 +21,26 @@ Terra.describeViewports('secondary scroll multiple directories', ['huge'], () =>
 
 Terra.describeViewports('secondary nav', ['huge'], () => {
   it('should not scroll when item is visible', () => {
-    browser.url('/secondary-nav-test/terra-dev-site/secondary-scroll-test/filler/filler-test');
+    browser.url('/secondary-nav-test/terra-dev-site/secondary-scroll-test/filler-02/filler-02-test');
     Terra.validates.element('should not scroll when item is visible before', { selector: '#root' });
-    browser.click('[class*="Extension-module__extension"]');
-    browser.keys('Filler 04');
-    browser.waitForVisible('[class*="list-item"]', 5000);
-    browser.click('[class*="list-item"]');
+    browser.click('[class*="filler-02"]');
+    browser.waitForVisible('[class*="filler-04"]', 5000);
     Terra.validates.element('should not scroll when item is visible after', { selector: '#root' });
   });
 
   it('should scroll up when item is not visible and at the top', () => {
     browser.url('/secondary-nav-test/terra-dev-site/secondary-scroll-test/filler-10/filler-10-test');
     Terra.validates.element('should scroll up when item is not visible and at the top before', { selector: '#root' });
-    browser.click('[class*="Extension-module__extension"]');
-    browser.keys('Filler 01');
-    browser.waitForVisible('[class*="list-item"]', 5000);
-    browser.click('[class*="list-item"]');
+    browser.click('[class*="filler-10"]');
+    browser.waitForVisible('[class*="filler-01"]', 5000);
     Terra.validates.element('should scroll up when item is not visible and at the top after', { selector: '#root' });
   });
 
   it('should scroll down when item is not visible and at the bottom', () => {
     browser.url('/secondary-nav-test/terra-dev-site/secondary-scroll-test/filler/filler-test');
     Terra.validates.element('should scroll down when item is not visible and at the bottom before', { selector: '#root' });
-    browser.click('[class*="Extension-module__extension"]');
-    browser.keys('Filler 17');
-    browser.waitForVisible('[class*="list-item"]', 5000);
-    browser.click('[class*="list-item"]');
+    browser.click('[class*="filler-00"]');
+    browser.waitForVisible('[class*="filler-17"]', 5000);
     Terra.validates.element('should scroll down when item is not visible and at the bottom after', { selector: '#root' });
   });